transaction_bundling/generic/python_example/pyproject.toml
[tool.poetry]
name = "transaction-bundling-generic-example"
version = "0.1.0"
description = "Example of generic transaction bundling"
authors = ["Compass Labs"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^7.12.0"
compass-api-sdk = "^0.9.17"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.16.0"
pytest = "^8.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
